Feat. HW0
Implement the "filteredByCycles" and "filteredByStreams" method to filter the food menu list.
Methods should only return food menu entries between "startTime" and "endTime".
The "UserMealWithExcess.excess" field should indicate whether the total calories for the whole day is greater than the "caloriesPerDay" value.
"UserMealWithExcess" is a record of one meal, but the "excess" field will be the same for all records for that day.